+++ mud-contracts/world/src/codegen/tables/SmartGateLinkTable.sol
@@ -16,26 +16,30 @@ import { Schema } from "@latticexyz/store/src/Schema.sol";
 import { EncodedLengths, EncodedLengthsLib } from "@latticexyz/store/src/EncodedLengths.sol";
 import { ResourceId } from "@latticexyz/store/src/ResourceId.sol";
 
+struct SmartGateLinkTableData {
+  uint256 destinationGateId;
+  bool isLinked;
+}
+
 library SmartGateLinkTable {
   // Hex below is the result of `WorldResourceIdLib.encode({ namespace: "eveworld", name: "SmartGateLinkTab", typeId: RESOURCE_TABLE });`
   ResourceId constant _tableId = ResourceId.wrap(0x7462657665776f726c64000000000000536d617274476174654c696e6b546162);
 
   FieldLayout constant _fieldLayout =
-    FieldLayout.wrap(0x0001010001000000000000000000000000000000000000000000000000000000);
+    FieldLayout.wrap(0x0021020020010000000000000000000000000000000000000000000000000000);
 
-  // Hex-encoded key schema of (uint256, uint256)
-  Schema constant _keySchema = Schema.wrap(0x004002001f1f0000000000000000000000000000000000000000000000000000);
-  // Hex-encoded value schema of (bool)
-  Schema constant _valueSchema = Schema.wrap(0x0001010060000000000000000000000000000000000000000000000000000000);
+  // Hex-encoded key schema of (uint256)
+  Schema constant _keySchema = Schema.wrap(0x002001001f000000000000000000000000000000000000000000000000000000);
+  // Hex-encoded value schema of (uint256, bool)
+  Schema constant _valueSchema = Schema.wrap(0x002102001f600000000000000000000000000000000000000000000000000000);
 
   /**
    * @notice Get the table's key field names.
    * @return keyNames An array of strings with the names of key fields.
    */
   function getKeyNames() internal pure returns (string[] memory keyNames) {
-    keyNames = new string[](2);
+    keyNames = new string[](1);
     keyNames[0] = "sourceGateId";
-    keyNames[1] = "destinationGateId";
   }
 
   /**
@@ -43,8 +47,9 @@ library SmartGateLinkTable {
    * @return fieldNames An array of strings with the names of value fields.
    */
   function getFieldNames() internal pure returns (string[] memory fieldNames) {
-    fieldNames = new string[](1);
-    fieldNames[0] = "isLinked";
+    fieldNames = new string[](2);
+    fieldNames[0] = "destinationGateId";
+    fieldNames[1] = "isLinked";
   }
 
   /**
@@ -62,104 +67,208 @@ library SmartGateLinkTable {
   }
 
   /**
-   * @notice Get isLinked.
+   * @notice Get destinationGateId.
    */
-  function getIsLinked(uint256 sourceGateId, uint256 destinationGateId) internal view returns (bool isLinked) {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function getDestinationGateId(uint256 sourceGateId) internal view returns (uint256 destinationGateId) {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
     bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
-    return (_toBool(uint8(bytes1(_blob))));
+    return (uint256(bytes32(_blob)));
   }
 
   /**
-   * @notice Get isLinked.
+   * @notice Get destinationGateId.
    */
-  function _getIsLinked(uint256 sourceGateId, uint256 destinationGateId) internal view returns (bool isLinked) {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function _getDestinationGateId(uint256 sourceGateId) internal view returns (uint256 destinationGateId) {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
     bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
+    return (uint256(bytes32(_blob)));
+  }
+
+  /**
+   * @notice Set destinationGateId.
+   */
+  function setDestinationGateId(uint256 sourceGateId, uint256 destinationGateId) internal {
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    StoreSwitch.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((destinationGateId)), _fieldLayout);
+  }
+
+  /**
+   * @notice Set destinationGateId.
+   */
+  function _setDestinationGateId(uint256 sourceGateId, uint256 destinationGateId) internal {
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    StoreCore.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((destinationGateId)), _fieldLayout);
+  }
+
+  /**
+   * @notice Get isLinked.
+   */
+  function getIsLinked(uint256 sourceGateId) internal view returns (bool isLinked) {
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 1, _fieldLayout);
     return (_toBool(uint8(bytes1(_blob))));
   }
 
   /**
    * @notice Get isLinked.
    */
-  function get(uint256 sourceGateId, uint256 destinationGateId) internal view returns (bool isLinked) {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function _getIsLinked(uint256 sourceGateId) internal view returns (bool isLinked) {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
-    bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
-    return (_toBool(uint8(bytes1(_blob))));
-  }
-
-  /**
-   * @notice Get isLinked.
-   */
-  function _get(uint256 sourceGateId, uint256 destinationGateId) internal view returns (bool isLinked) {
-    bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
-
-    bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
+    bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 1, _fieldLayout);
     return (_toBool(uint8(bytes1(_blob))));
   }
 
   /**
    * @notice Set isLinked.
    */
-  function setIsLinked(uint256 sourceGateId, uint256 destinationGateId, bool isLinked) internal {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function setIsLinked(uint256 sourceGateId, bool isLinked) internal {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
-    StoreSwitch.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((isLinked)), _fieldLayout);
+    StoreSwitch.setStaticField(_tableId, _keyTuple, 1, abi.encodePacked((isLinked)), _fieldLayout);
   }
 
   /**
    * @notice Set isLinked.
    */
-  function _setIsLinked(uint256 sourceGateId, uint256 destinationGateId, bool isLinked) internal {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function _setIsLinked(uint256 sourceGateId, bool isLinked) internal {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
-    StoreCore.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((isLinked)), _fieldLayout);
+    StoreCore.setStaticField(_tableId, _keyTuple, 1, abi.encodePacked((isLinked)), _fieldLayout);
   }
 
   /**
-   * @notice Set isLinked.
+   * @notice Get the full data.
+   */
+  function get(uint256 sourceGateId) internal view returns (SmartGateLinkTableData memory _table) {
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    (bytes memory _staticData, EncodedLengths _encodedLengths, bytes memory _dynamicData) = StoreSwitch.getRecord(
+      _tableId,
+      _keyTuple,
+      _fieldLayout
+    );
+    return decode(_staticData, _encodedLengths, _dynamicData);
+  }
+
+  /**
+   * @notice Get the full data.
+   */
+  function _get(uint256 sourceGateId) internal view returns (SmartGateLinkTableData memory _table) {
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    (bytes memory _staticData, EncodedLengths _encodedLengths, bytes memory _dynamicData) = StoreCore.getRecord(
+      _tableId,
+      _keyTuple,
+      _fieldLayout
+    );
+    return decode(_staticData, _encodedLengths, _dynamicData);
+  }
+
+  /**
+   * @notice Set the full data using individual values.
    */
   function set(uint256 sourceGateId, uint256 destinationGateId, bool isLinked) internal {
-    bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
+    bytes memory _staticData = encodeStatic(destinationGateId, isLinked);
 
-    StoreSwitch.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((isLinked)), _fieldLayout);
+    EncodedLengths _encodedLengths;
+    bytes memory _dynamicData;
+
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    StoreSwitch.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData);
   }
 
   /**
-   * @notice Set isLinked.
+   * @notice Set the full data using individual values.
    */
   function _set(uint256 sourceGateId, uint256 destinationGateId, bool isLinked) internal {
-    bytes32[] memory _keyTuple = new bytes32[](2);
-    _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
+    bytes memory _staticData = encodeStatic(destinationGateId, isLinked);
 
-    StoreCore.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked((isLinked)), _fieldLayout);
+    EncodedLengths _encodedLengths;
+    bytes memory _dynamicData;
+
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    StoreCore.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData, _fieldLayout);
+  }
+
+  /**
+   * @notice Set the full data using the data struct.
+   */
+  function set(uint256 sourceGateId, SmartGateLinkTableData memory _table) internal {
+    bytes memory _staticData = encodeStatic(_table.destinationGateId, _table.isLinked);
+
+    EncodedLengths _encodedLengths;
+    bytes memory _dynamicData;
+
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    StoreSwitch.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData);
+  }
+
+  /**
+   * @notice Set the full data using the data struct.
+   */
+  function _set(uint256 sourceGateId, SmartGateLinkTableData memory _table) internal {
+    bytes memory _staticData = encodeStatic(_table.destinationGateId, _table.isLinked);
+
+    EncodedLengths _encodedLengths;
+    bytes memory _dynamicData;
+
+    bytes32[] memory _keyTuple = new bytes32[](1);
+    _keyTuple[0] = bytes32(uint256(sourceGateId));
+
+    StoreCore.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData, _fieldLayout);
+  }
+
+  /**
+   * @notice Decode the tightly packed blob of static data using this table's field layout.
+   */
+  function decodeStatic(bytes memory _blob) internal pure returns (uint256 destinationGateId, bool isLinked) {
+    destinationGateId = (uint256(Bytes.getBytes32(_blob, 0)));
+
+    isLinked = (_toBool(uint8(Bytes.getBytes1(_blob, 32))));
+  }
+
+  /**
+   * @notice Decode the tightly packed blobs using this table's field layout.
+   * @param _staticData Tightly packed static fields.
+   *
+   *
+   */
+  function decode(
+    bytes memory _staticData,
+    EncodedLengths,
+    bytes memory
+  ) internal pure returns (SmartGateLinkTableData memory _table) {
+    (_table.destinationGateId, _table.isLinked) = decodeStatic(_staticData);
   }
 
   /**
    * @notice Delete all data for given keys.
    */
-  function deleteRecord(uint256 sourceGateId, uint256 destinationGateId) internal {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function deleteRecord(uint256 sourceGateId) internal {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
     StoreSwitch.deleteRecord(_tableId, _keyTuple);
   }
@@ -167,10 +276,9 @@ library SmartGateLinkTable {
   /**
    * @notice Delete all data for given keys.
    */
-  function _deleteRecord(uint256 sourceGateId, uint256 destinationGateId) internal {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function _deleteRecord(uint256 sourceGateId) internal {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
     StoreCore.deleteRecord(_tableId, _keyTuple, _fieldLayout);
   }
@@ -179,8 +287,8 @@ library SmartGateLinkTable {
    * @notice Tightly pack static (fixed length) data using this table's schema.
    * @return The static data, encoded into a sequence of bytes.
    */
-  function encodeStatic(bool isLinked) internal pure returns (bytes memory) {
-    return abi.encodePacked(isLinked);
+  function encodeStatic(uint256 destinationGateId, bool isLinked) internal pure returns (bytes memory) {
+    return abi.encodePacked(destinationGateId, isLinked);
   }
 
   /**
@@ -189,8 +297,11 @@ library SmartGateLinkTable {
    * @return The lengths of the dynamic fields (packed into a single bytes32 value).
    * @return The dynamic (variable length) data, encoded into a sequence of bytes.
    */
-  function encode(bool isLinked) internal pure returns (bytes memory, EncodedLengths, bytes memory) {
-    bytes memory _staticData = encodeStatic(isLinked);
+  function encode(
+    uint256 destinationGateId,
+    bool isLinked
+  ) internal pure returns (bytes memory, EncodedLengths, bytes memory) {
+    bytes memory _staticData = encodeStatic(destinationGateId, isLinked);
 
     EncodedLengths _encodedLengths;
     bytes memory _dynamicData;
@@ -201,10 +312,9 @@ library SmartGateLinkTable {
   /**
    * @notice Encode keys as a bytes32 array using this table's field layout.
    */
-  function encodeKeyTuple(uint256 sourceGateId, uint256 destinationGateId) internal pure returns (bytes32[] memory) {
-    bytes32[] memory _keyTuple = new bytes32[](2);
+  function encodeKeyTuple(uint256 sourceGateId) internal pure returns (bytes32[] memory) {
+    bytes32[] memory _keyTuple = new bytes32[](1);
     _keyTuple[0] = bytes32(uint256(sourceGateId));
-    _keyTuple[1] = bytes32(uint256(destinationGateId));
 
     return _keyTuple;
   }