+++ mud-contracts/world/src/codegen/world/ISmartDeployableSystem.sol
@@ -18,7 +18,7 @@ interface ISmartDeployableSystem {
     uint256 entityId,
     SmartObjectData memory smartObjectData,
     uint256 fuelUnitVolumeInWei,
-    uint256 fuelConsumptionPerMinuteInWei,
+    uint256 fuelConsumptionIntervalInSeconds,
     uint256 fuelMaxCapacityInWei
   ) external;
 
@@ -38,7 +38,7 @@ interface ISmartDeployableSystem {
 
   function eveworld__globalResume() external;
 
-  function eveworld__setFuelConsumptionPerMinute(uint256 entityId, uint256 fuelConsumptionPerMinuteInWei) external;
+  function eveworld__setFuelConsumptionPerMinute(uint256 entityId, uint256 fuelConsumptionIntervalInSeconds) external;
 
   function eveworld__setFuelMaxCapacity(uint256 entityId, uint256 capacityInWei) external;