+++ mud-contracts/world/src/codegen/world/ISmartTurretSystem.sol
@@ -31,6 +31,7 @@ interface ISmartTurretSystem {
 
   function eveworld__inProximity(
     uint256 smartTurretId,
+    uint256 turretOwnerCharacterId,
     TargetPriority[] memory priorityQueue,
     Turret memory turret,
     SmartTurretTarget memory turretTarget
@@ -38,6 +39,7 @@ interface ISmartTurretSystem {
 
   function eveworld__aggression(
     uint256 smartTurretId,
+    uint256 turretOwnerCharacterId,
     TargetPriority[] memory priorityQueue,
     Turret memory turret,
     SmartTurretTarget memory aggressor,