+++ mud-contracts/world/src/modules/smart-storage-unit/systems/SmartStorageUnitSystem.sol
@@ -48,7 +48,7 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
   /**
    * @notice Create and anchor a smart storage unit
    * @dev Create and anchor a smart storage unit by smart object id
-   * @param smartObjectId The smart object id
+   * @param smartStorageUnitId The smart object id
    * @param entityRecordData The entity record data of the smart object
    * @param smartObjectData is the token metadata of the smart object
    * @param worldPosition The position of the smart object in the game
@@ -56,16 +56,16 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
    * @param ephemeralStorageCapacity The personal storage capacity of the smart storage unit
    */
   function createAndAnchorSmartStorageUnit(
-    uint256 smartObjectId,
+    uint256 smartStorageUnitId,
     EntityRecordData memory entityRecordData,
     SmartObjectData memory smartObjectData,
     WorldPosition memory worldPosition,
     uint256 fuelUnitVolume,
-    uint256 fuelConsumptionPerMinute,
+    uint256 fuelConsumptionIntervalInSeconds,
     uint256 fuelMaxCapacity,
     uint256 storageCapacity,
     uint256 ephemeralStorageCapacity
-  ) public onlyAdmin hookable(smartObjectId, _systemId()) {
+  ) public onlyAdmin hookable(smartStorageUnitId, _systemId()) {
     {
       uint256 classId = ClassConfig.getClassId(_systemId());
 
@@ -73,29 +73,29 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
         revert ISmartStorageUnitErrors.SmartStorageUnit_UndefinedClassId();
       }
 
-      if (EntityTable.getDoesExists(smartObjectId) == false) {
-        // register smartObjectId as an object
-        _smartObjectLib().registerEntity(smartObjectId, OBJECT);
+      if (EntityTable.getDoesExists(smartStorageUnitId) == false) {
+        // register smartStorageUnitId as an object
+        _smartObjectLib().registerEntity(smartStorageUnitId, OBJECT);
         // tag this object's entity Id to a set of defined classIds
-        _smartObjectLib().tagEntity(smartObjectId, classId);
+        _smartObjectLib().tagEntity(smartStorageUnitId, classId);
       }
     }
     //Implement the logic to store the data in different modules: EntityRecord, Deployable, Location and ERC721
     _entityRecordLib().createEntityRecord(
-      smartObjectId,
+      smartStorageUnitId,
       entityRecordData.itemId,
       entityRecordData.typeId,
       entityRecordData.volume
     );
 
     _smartDeployableLib().registerDeployable(
-      smartObjectId,
+      smartStorageUnitId,
       smartObjectData,
       fuelUnitVolume,
-      fuelConsumptionPerMinute,
+      fuelConsumptionIntervalInSeconds,
       fuelMaxCapacity
     );
-    _smartDeployableLib().setSmartAssemblyType(smartObjectId, SmartAssemblyType.SMART_STORAGE_UNIT);
+    _smartDeployableLib().setSmartAssemblyType(smartStorageUnitId, SmartAssemblyType.SMART_STORAGE_UNIT);
 
     LocationTableData memory locationData = LocationTableData({
       solarSystemId: worldPosition.solarSystemId,
@@ -103,10 +103,10 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
       y: worldPosition.position.y,
       z: worldPosition.position.z
     });
-    _smartDeployableLib().anchor(smartObjectId, locationData);
+    _smartDeployableLib().anchor(smartStorageUnitId, locationData);
 
-    _inventoryLib().setInventoryCapacity(smartObjectId, storageCapacity);
-    _inventoryLib().setEphemeralInventoryCapacity(smartObjectId, ephemeralStorageCapacity);
+    _inventoryLib().setInventoryCapacity(smartStorageUnitId, storageCapacity);
+    _inventoryLib().setEphemeralInventoryCapacity(smartStorageUnitId, ephemeralStorageCapacity);
   }
 
   /**
@@ -114,13 +114,13 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
    * @dev Create an item by smart object id and deposit it to the inventory
    * //TODO only server account can create items on-chain
    * //TODO Represent item as a ERC1155 asset with ownership on-chain
-   * @param smartObjectId The smart object id
+   * @param smartStorageUnitId The smart object id
    * @param items The item to store in a inventory
    */
   function createAndDepositItemsToInventory(
-    uint256 smartObjectId,
+    uint256 smartStorageUnitId,
     InventoryItem[] memory items
-  ) public onlyAdmin hookable(smartObjectId, _systemId()) {
+  ) public onlyAdmin hookable(smartStorageUnitId, _systemId()) {
     for (uint256 i = 0; i < items.length; i++) {
       //Check if the item exists on-chain if not Create entityRecord
       _entityRecordLib().createEntityRecord(
@@ -131,7 +131,7 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
       );
     }
     //Deposit item to the inventory
-    _inventoryLib().depositToInventory(smartObjectId, items);
+    _inventoryLib().depositToInventory(smartStorageUnitId, items);
   }
 
   /**
@@ -139,15 +139,15 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
    * @dev Create an item by smart object id and deposit it to the ephemeral inventory
    * //TODO only server account can create items on-chain
    * //TODO Represent item as a ERC1155 asset with ownership on-chain
-   * @param smartObjectId The smart object id
+   * @param smartStorageUnitId The smart object id
    * @param ephemeralInventoryOwner The owner of the inventory
    * @param items The item to store in a inventory
    */
   function createAndDepositItemsToEphemeralInventory(
-    uint256 smartObjectId,
+    uint256 smartStorageUnitId,
     address ephemeralInventoryOwner,
     InventoryItem[] memory items
-  ) public onlyAdmin hookable(smartObjectId, _systemId()) {
+  ) public onlyAdmin hookable(smartStorageUnitId, _systemId()) {
     //Check if the item exists on-chain if not Create entityRecord
     for (uint256 i = 0; i < items.length; i++) {
       _entityRecordLib().createEntityRecord(
@@ -159,7 +159,7 @@ contract SmartStorageUnitSystem is AccessModified, EveSystem {
     }
     //Deposit item to the ephemeral inventory
     // TODO: This _might_ clash with online fuel, since that would require the underlying deployable to be funded in fuel
-    _inventoryLib().depositToEphemeralInventory(smartObjectId, ephemeralInventoryOwner, items);
+    _inventoryLib().depositToEphemeralInventory(smartStorageUnitId, ephemeralInventoryOwner, items);
   }
 
   function setSSUClassId(