+++ mud-contracts/world/src/modules/smart-turret/systems/SmartTurretSystem.sol
@@ -17,6 +17,7 @@ import { DeployableState, DeployableStateData } from "../../../codegen/tables/De
 import { LocationTableData } from "../../../codegen/tables/LocationTable.sol";
 import { ClassConfig } from "../../../codegen/tables/ClassConfig.sol";
 import { State, SmartAssemblyType } from "../../../codegen/common.sol";
+import { CharactersTable } from "../../../codegen/tables/CharactersTable.sol";
 
 import { EntityRecordData, WorldPosition } from "../../smart-storage-unit/types.sol";
 import { EntityRecordLib } from "../../entity-record/EntityRecordLib.sol";
@@ -26,6 +27,7 @@ import { SmartDeployableLib } from "../../smart-deployable/SmartDeployableLib.so
 import { SmartDeployableLib } from "../../smart-deployable/SmartDeployableLib.sol";
 import { SmartObjectData } from "../../smart-deployable/types.sol";
 import { Utils as SmartDeployableUtils } from "../../smart-deployable/Utils.sol";
+import { Utils as SmartCharacterUtils } from "../../smart-character/Utils.sol";
 import { AccessModified } from "../../access/systems/AccessModified.sol";
 
 import { Utils } from "../Utils.sol";
@@ -42,6 +44,7 @@ contract SmartTurretSystem is EveSystem, AccessModified {
   using SmartDeployableLib for SmartDeployableLib.World;
   using SmartObjectFrameworkUtils for bytes14;
   using SmartDeployableUtils for bytes14;
+  using SmartCharacterUtils for bytes14;
   using Utils for bytes14;
 
   error SmartTurret_UndefinedClassId();
@@ -121,12 +124,14 @@ contract SmartTurretSystem is EveSystem, AccessModified {
   /**
    * @notice view function for turret logic based on proximity
    * @param smartTurretId is the is of the smart turret
+   * @param turretOwnerCharacterId is the character id of the owner of the smart turret
    * @param priorityQueue is the queue of the SmartTurretTarget in proximity
    * @param turret is the Smart Turret object
    * @param turretTarget is the player entering the zone
    */
   function inProximity(
     uint256 smartTurretId,
+    uint256 turretOwnerCharacterId,
     TargetPriority[] memory priorityQueue,
     Turret memory turret,
     SmartTurretTarget memory turretTarget
@@ -139,14 +144,26 @@ contract SmartTurretSystem is EveSystem, AccessModified {
     // Delegate the call to the implementation inProximity view function
     ResourceId systemId = SmartTurretConfigTable.get(smartTurretId);
 
+    //If smart turret is not configured, then execute the default logic
     if (!ResourceIds.getExists(systemId)) {
-      //If smart turret is not configured, then execute the default logic
-      // TODO: If the character corp and the owner of the turret are same, then the turret will not attack
-      updatedPriorityQueue = priorityQueue; //temporary logic
+      //If the corp and the smart turret owner of the target turret are same, then the turret will not attack
+      uint256 smartTurretOwnerCorp = CharactersTable.getCorpId(turretOwnerCharacterId);
+      uint256 turretTargetCorp = CharactersTable.getCorpId(turretTarget.characterId);
+      if (smartTurretOwnerCorp != turretTargetCorp) {
+        updatedPriorityQueue = new TargetPriority[](priorityQueue.length + 1);
+        for (uint256 i = 0; i < priorityQueue.length; i++) {
+          updatedPriorityQueue[i] = priorityQueue[i];
+        }
+
+        updatedPriorityQueue[priorityQueue.length] = TargetPriority({ target: turretTarget, weight: 1 }); //should the weight be 1? or the heighest of all weights in the array ?
+      } else {
+        //If the corp and the smart turret owner of the target turret are same, then do not add the target turret to the priority queue
+        updatedPriorityQueue = priorityQueue;
+      }
     } else {
       bytes memory returnData = world().call(
         systemId,
-        abi.encodeCall(this.inProximity, (smartTurretId, priorityQueue, turret, turretTarget))
+        abi.encodeCall(this.inProximity, (smartTurretId, turretOwnerCharacterId, priorityQueue, turret, turretTarget))
       );
 
       updatedPriorityQueue = abi.decode(returnData, (TargetPriority[]));
@@ -157,6 +174,7 @@ contract SmartTurretSystem is EveSystem, AccessModified {
 
   /**
    * @param smartTurretId is the is of the smart turret
+   * @param turretOwnerCharacterId is the character id of the owner of the smart turret
    * @param priorityQueue is the queue of the SmartTurretTarget in proximity
    * @param turret is the Smart Turret object
    * @param aggressor is the player attacking inside the zone
@@ -164,6 +182,7 @@ contract SmartTurretSystem is EveSystem, AccessModified {
    */
   function aggression(
     uint256 smartTurretId,
+    uint256 turretOwnerCharacterId,
     TargetPriority[] memory priorityQueue,
     Turret memory turret,
     SmartTurretTarget memory aggressor,
@@ -178,12 +197,25 @@ contract SmartTurretSystem is EveSystem, AccessModified {
     ResourceId systemId = SmartTurretConfigTable.get(smartTurretId);
 
     if (!ResourceIds.getExists(systemId)) {
-      //If smart turret is not configured, then execute the default logic
-      updatedPriorityQueue = priorityQueue; //temporary logic
+      //If the corp of the smart turret owner of the aggressor are same, then the turret will not attack
+      uint256 turretOwnerCorp = CharactersTable.getCorpId(turretOwnerCharacterId);
+      uint256 aggressorCorp = CharactersTable.getCorpId(aggressor.characterId);
+
+      if (turretOwnerCorp != aggressorCorp) {
+        updatedPriorityQueue = new TargetPriority[](priorityQueue.length + 1);
+        for (uint256 i = 0; i < priorityQueue.length; i++) {
+          updatedPriorityQueue[i] = priorityQueue[i];
+        }
+
+        updatedPriorityQueue[priorityQueue.length] = TargetPriority({ target: aggressor, weight: 1 }); //should the weight be 1? or the heighest of all weights in the array ?
+      }
     } else {
       bytes memory returnData = world().call(
         systemId,
-        abi.encodeCall(this.aggression, (smartTurretId, priorityQueue, turret, aggressor, victim))
+        abi.encodeCall(
+          this.aggression,
+          (smartTurretId, turretOwnerCharacterId, priorityQueue, turret, aggressor, victim)
+        )
       );
 
       updatedPriorityQueue = abi.decode(returnData, (TargetPriority[]));