+++ mud-contracts/world/test/inventory/Interact.t.sol
@@ -96,7 +96,7 @@ contract InteractTest is MudTest {
       smartObjectData,
       worldPosition,
       1e18, // fuelUnitVolume,
-      1, // fuelConsumptionPerMinute,
+      1, // fuelConsumptionIntervalInSeconds,
       1000000 * 1e18, // fuelMaxCapacity,
       storageCapacity,
       ephemeralStorageCapacity